LayoutTests/media/session/track-media-events-disabled-expected.txt

Previous/Next Track events should not be fired if previous/nextTrackEnabled are 'false'.


EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
Begin playing media.
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
Sending Next Track media event.
RUN(internals.sendMediaControlEvent("next-track"))
Sending Previous Track media event.
RUN(internals.sendMediaControlEvent("previous-track"))
END OF TEST
